[DO NOT MERGE] [WIP] Push so we can test against integration squidex
