+++ packages/validation/src/index.ts
@@ -1,3 +1,2 @@
 export * from './user';
 export * from './teams';
-export { VALID_ORCID } from './validate-orcids';
