+++ packages/validation/src/user.ts
@@ -3,6 +3,7 @@ import { UserResponse } from '@asap-hub/model';
 export const USER_SOCIAL_WEBSITE = /^http(s?):\/\/\S+/i;
 export const USER_SOCIAL_RESEARCHER_ID = /^[a-z]{1,3}-\d{4}-20\d\d$/i;
 export const USER_SOCIAL_NOT_URL = /^(?!\s*http(s?):\/\/)\S+/i;
+export const VALID_ORCID = /^\d{4}-\d{4}-\d{4}-\d{3}(\d|X)$/;
 
 const getFailedResponse = () => ({
   isOnboardable: false,
